# [SPARK-48922][SQL] Optimize nested data type insertion performance