<!--
Thanks for sending a pull request!  Here are some tips for you:
  1. If this is your first time, please read our contributor guidelines: https://spark.apache.org/contributing.html
  2. Ensure you have added or run the appropriate tests for your PR: https://spark.apache.org/developer-tools.html
  3. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP][SPARK-XXXX] Your PR title ...'.
  4. Be sure to keep the PR description updated to reflect all changes.
  5. Please write your PR title to summarize what this PR proposes.
  6. If possible, provide a concise example to reproduce the issue for a faster review.
  7. If you want to add a new configuration, please read the guideline first for naming configurations in
     'core/src/main/scala/org/apache/spark/internal/config/ConfigEntry.scala'.
  8. If you want to add or modify an error type or message, please read the guideline first in
     'common/utils/src/main/resources/error/README.md'.
-->

### What changes were proposed in this pull request?

To improve insertion performance, we do not need to add transform expressions when there is no conversion for complex types.

### Why are the changes needed?
During the upgrade from Spark 3.1.1 to 3.5.0, I found a performance regression in map type inserts.

There are some extra conversion expressions in project before insert, which doesn't seem to be always necessary.

```
map_from_arrays(transform(map_keys(map#516), lambdafunction(lambda key#652, lambda key#652, false)), transform(map_values(map#516), lambdafunction(lambda value#654, lambda value#654, false))) AS map#656 
```

### Does this PR introduce _any_ user-facing change?

No


### How was this patch tested?
<!--
If tests were added, say they were added here. Please make sure to add some test cases that check the changes thoroughly including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please clarify how you tested step by step, ideally copy and paste-able, so that other reviewers can test and check, and descendants can verify in the future.
If tests were not added, please describe why they were not added and/or why it was difficult to add.
If benchmark tests were added, please run the benchmarks in GitHub Actions for the consistent environment, and the instructions could accord to: https://spark.apache.org/developer-tools.html#github-workflow-benchmarks.
-->
Added benchmark case.

before this pr:
```
Java HotSpot(TM) 64-Bit Server VM 17.0.10+11-LTS-240 on Linux 3.10.0-957.el7.x86_64
Intel Core Processor (Skylake, IBRS)
Insert map type table benchmark, totalRows = 10240:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
----------------------------------------------------------------------------------------------------------------------------------
insert map type table with conversion                        1054           1092          54          0.0      102906.7       1.0X
insert map type table without conversion                      540            582          47          0.0       52770.3       2.0X

Java HotSpot(TM) 64-Bit Server VM 17.0.10+11-LTS-240 on Linux 3.10.0-957.el7.x86_64
Intel Core Processor (Skylake, IBRS)
Insert array type table benchmark, totalRows = 10240:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------------------
insert array type table with conversion                         372            399          20          0.0       36337.9       1.0X
insert array type table without conversion                      233            255          19          0.0       22775.5       1.6X
```

after this pr:
```
Java HotSpot(TM) 64-Bit Server VM 17.0.10+11-LTS-240 on Linux 3.10.0-957.el7.x86_64
Intel Core Processor (Skylake, IBRS)
Insert map type table benchmark, totalRows = 10240:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
----------------------------------------------------------------------------------------------------------------------------------
insert map type table with conversion                        1222           1226           6          0.0      119360.0       1.0X
insert map type table without conversion                      417            429          12          0.0       40714.5       2.9X

Java HotSpot(TM) 64-Bit Server VM 17.0.10+11-LTS-240 on Linux 3.10.0-957.el7.x86_64
Intel Core Processor (Skylake, IBRS)
Insert array type table benchmark, totalRows = 10240:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------------------
insert array type table with conversion                         389            433          31          0.0       37967.2       1.0X
insert array type table without conversion                      207            241          23          0.0       20246.6       1.9X
```

### Was this patch authored or co-authored using generative AI tooling?
<!--
If generative AI tooling has been used in the process of authoring this patch, please include the
phrase: 'Generated-by: ' followed by the name of the tool and its version.
If no, write 'No'.
Please refer to the [ASF Generative Tooling Guidance](https://www.apache.org/legal/generative-tooling.html) for details.
-->
No