@wForget 2024-07-18
The previous behavior was introduced in #33728, @cloud-fan could you please take a look?