@pan3793 2024-10-11
Is the fix proposed here partially fixed by SPARK-49352? also cc @viirya