@wForget 2024-10-11
> Is the fix proposed here partially fixed by [SPARK-49352](https://issues.apache.org/jira/browse/SPARK-49352)? also cc @viirya


 [SPARK-49352](https://issues.apache.org/jira/browse/SPARK-49352) does not seem to cover `MapType`