@kazuyukitanimura 2025-03-06
Hi @wForget 
This PR looks important, do you plan to reopen this and rebase on top of [SPARK-49352](https://issues.apache.org/jira/browse/SPARK-49352) ?
https://github.com/apache/spark/pull/47843