@wForget 2025-03-07
> Hi @wForget This PR looks important, do you plan to reopen this and rebase on top of [SPARK-49352](https://issues.apache.org/jira/browse/SPARK-49352) ? #47843

Sure, I will rebase on master and submit a new pr.